refactor: improve object deletion logic and streamline group association handling in well inventory processing

+++ tests/features/environment.py
@@ -509,7 +509,9 @@ def after_all(context):
     with session_ctx() as session:
         for table in context.objects.values():
             for obj in table:
-                session.delete(obj)
+                obj = session.get(type(obj), obj.id)
+                if obj:
+                    session.delete(obj)
 
         # session.query(TransducerObservationBlock).delete()
         # session.query(TransducerObservation).delete()
